Support css-loader v2
css-loader version 2 uses a placeholder constant and a single require
for url resources. Something about this causes the dependency resolution
to only be called once. This means that there will be more instances of
the placeholder string in the content than are in the
extractedDependencyContent array. This is fixed by explicitly tying
the dependencies to a unique instance of the placeholder string.