feat: Add support for ECMAScript modules (#69)
* Adds support for the new file-loader which emits ESMs by default
* Uses Babel to transpile the source code
* Update test to handle the new output from babel
* Update package-lock.json

BREAKING CHANGE: The extract-loader now uses Babel to transpile the bundle code for the current Node version. This is required in order to support the new file-loader which produces ECMAScript modules. This *should not* be a breaking change but since Babel is also known to be kind of brittle in regards of configuration we cannot guarantee that there might be unknown/unwanted side-effects in your project setup.